requirements.txt
# openai
fastapi
uvicorn
httpx
beautifulsoup4
pydantic
python-dotenv
google-search-results
groq